has problem of converting from uintptr to float64

+++ cv/videoio.go
@@ -40,6 +40,6 @@ func (cap *VideoCapture) IsOpened() bool {
 	return cvVideoCaptureIsOpened(cap)
 }
 
-func (cap *VideoCapture) Get(propId int) float64 {
+func (cap *VideoCapture) Get(propId int) int32 {
 	return cvVideoCaptureGet(cap, propId)
 }
